move pallet asset code logic to utils -> pallet_assets.rs file
