+++ contracts/token/psp22/mod.rs
@@ -33,4 +33,5 @@ pub mod extensions {
 
 pub mod utils {
     pub mod token_timelock;
+    pub mod pallet_assets;
 }
