+++ examples/psp22_pallet_asset/lib.rs
@@ -15,281 +15,13 @@ use brush::{
             metadata::*,
             mintable::*,
         },
+        utils::*,
         PSP22Error,
     },
     modifiers,
 };
 use ink_lang::ChainExtensionInstance;
-pub struct PalletAsset;
-
-impl PalletAsset {
-    fn create(
-        origin_type: OriginType,
-        asset_id: u32,
-        target_address: [u8; 32],
-        min_balance: u128,
-    ) -> Result<(), PalletAssetErr> {
-        let subject = PalletAssetRequest {
-            origin_type,
-            asset_id,
-            target_address,
-            amount: min_balance,
-        };
-        ::ink_env::chain_extension::ChainExtensionMethod::build(1102u32)
-            .input::<PalletAssetRequest>()
-            .output::<Result<(), PalletAssetErr>>()
-            .handle_error_code::<PalletAssetErr>()
-            .call(&subject)?
-    }
-
-    fn mint(
-        origin_type: OriginType,
-        asset_id: u32,
-        target_address: [u8; 32],
-        amount: u128,
-    ) -> Result<(), PalletAssetErr> {
-        let subject = PalletAssetRequest {
-            origin_type,
-            asset_id,
-            target_address,
-            amount,
-        };
-        ::ink_env::chain_extension::ChainExtensionMethod::build(1103u32)
-            .input::<PalletAssetRequest>()
-            .output::<Result<(), PalletAssetErr>>()
-            .handle_error_code::<PalletAssetErr>()
-            .call(&subject)?
-    }
-
-    fn burn(
-        origin_type: OriginType,
-        asset_id: u32,
-        target_address: [u8; 32],
-        amount: u128,
-    ) -> Result<(), PalletAssetErr> {
-        let subject = PalletAssetRequest {
-            origin_type,
-            asset_id,
-            target_address,
-            amount,
-        };
-        ::ink_env::chain_extension::ChainExtensionMethod::build(1104u32)
-            .input::<PalletAssetRequest>()
-            .output::<Result<(), PalletAssetErr>>()
-            .handle_error_code::<PalletAssetErr>()
-            .call(&subject)?
-    }
-
-    fn transfer(
-        origin_type: OriginType,
-        asset_id: u32,
-        target_address: [u8; 32],
-        amount: u128,
-    ) -> Result<(), PalletAssetErr> {
-        let subject = PalletAssetRequest {
-            origin_type,
-            asset_id,
-            target_address,
-            amount,
-        };
-        ::ink_env::chain_extension::ChainExtensionMethod::build(1105u32)
-            .input::<PalletAssetRequest>()
-            .output::<Result<(), PalletAssetErr>>()
-            .handle_error_code::<PalletAssetErr>()
-            .call(&subject)?
-    }
-
-    fn balance(asset_id: u32, address: [u8; 32]) -> Result<u128, PalletAssetErr> {
-        let subject = PalletAssetBalanceRequest { asset_id, address };
-        ::ink_env::chain_extension::ChainExtensionMethod::build(1106u32)
-            .input::<PalletAssetBalanceRequest>()
-            .output::<u128>()
-            .handle_error_code::<PalletAssetErr>()
-            .call(&subject)
-    }
-
-	fn total_supply(asset_id: u32) -> Result<u128, PalletAssetErr> {
-        ::ink_env::chain_extension::ChainExtensionMethod::build(1107u32)
-            .input::<u32>()
-            .output::<u128>()
-			.handle_error_code::<PalletAssetErr>()
-            .call(&asset_id)
-    }
-
-	fn approve_transfer(
-        origin_type: OriginType,
-        asset_id: u32,
-        target_address: [u8; 32],
-        amount: u128,
-    ) -> Result<(), PalletAssetErr> {
-        let subject = PalletAssetRequest {
-            origin_type,
-            asset_id,
-            target_address,
-            amount,
-        };
-        ::ink_env::chain_extension::ChainExtensionMethod::build(1108u32)
-            .input::<PalletAssetRequest>()
-            .output::<Result<(), PalletAssetErr>>()
-            .handle_error_code::<PalletAssetErr>()
-            .call(&subject)?
-    }
-
-	fn transfer_approved(
-        origin_type: OriginType,
-        asset_id: u32,
-		owner: [u8; 32],
-        target_address: [u8; 32],
-        amount: u128,
-    ) -> Result<(), PalletAssetErr> {
-        let subject = PalletAssetRequest {
-            origin_type,
-            asset_id,
-            target_address,
-            amount,
-        };
-        ::ink_env::chain_extension::ChainExtensionMethod::build(1109u32)
-            .input::<([u8; 32], PalletAssetRequest)>()
-            .output::<Result<(), PalletAssetErr>>()
-            .handle_error_code::<PalletAssetErr>()
-            .call(&(owner, subject))?
-    }
-	
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, scale::Encode, scale::Decode)]
-#[cfg_attr(feature = "std", derive(scale_info::TypeInfo))]
-pub enum RequestType {
-    Create,
-    Mint,
-    Burn,
-    Transfer,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, scale::Encode, scale::Decode)]
-#[cfg_attr(feature = "std", derive(scale_info::TypeInfo))]
-pub enum OriginType {
-    Caller,
-    Address,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, scale::Encode, scale::Decode)]
-#[cfg_attr(feature = "std", derive(scale_info::TypeInfo))]
-pub struct PalletAssetRequest {
-    origin_type: OriginType,
-    asset_id: u32,
-    target_address: [u8; 32],
-    amount: u128,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, scale::Encode, scale::Decode)]
-#[cfg_attr(feature = "std", derive(scale_info::TypeInfo))]
-pub struct PalletAssetBalanceRequest {
-    asset_id: u32,
-    address: [u8; 32],
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, scale::Encode, scale::Decode)]
-#[cfg_attr(feature = "std", derive(scale_info::TypeInfo))]
-pub enum PalletAssetErr {
-    /// Some error occurred.
-    Other,
-    /// Failed to lookup some data.
-    CannotLookup,
-    /// A bad origin.
-    BadOrigin,
-    /// A custom error in a module.
-    Module,
-    /// At least one consumer is remaining so the account cannot be destroyed.
-    ConsumerRemaining,
-    /// There are no providers so the account cannot be created.
-    NoProviders,
-    /// There are too many consumers so the account cannot be created.
-    TooManyConsumers,
-    /// An error to do with tokens.
-    Token(PalletAssetTokenErr),
-    /// An arithmetic error.
-    Arithmetic(PalletAssetArithmeticErr),
-    // unknown error
-    Unknown,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, scale::Encode, scale::Decode)]
-#[cfg_attr(feature = "std", derive(scale_info::TypeInfo))]
-pub enum PalletAssetArithmeticErr {
-    /// Underflow.
-    Underflow,
-    /// Overflow.
-    Overflow,
-    /// Division by zero.
-    DivisionByZero,
-    // unknown error
-    Unknown,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq, scale::Encode, scale::Decode)]
-#[cfg_attr(feature = "std", derive(scale_info::TypeInfo))]
-pub enum PalletAssetTokenErr {
-    /// Funds are unavailable.
-    NoFunds,
-    /// Account that must exist would die.
-    WouldDie,
-    /// Account cannot exist with the funds that would be given.
-    BelowMinimum,
-    /// Account cannot be created.
-    CannotCreate,
-    /// The asset in question is unknown.
-    UnknownAsset,
-    /// Funds exist but are frozen.
-    Frozen,
-    /// Operation is not supported by the asset.
-    Unsupported,
-    // unknown error
-    Unknown,
-}
-
-impl From<u8> for OriginType{
-	fn from(origin: u8) -> OriginType {
-		if origin == 0 {
-			OriginType::Caller
-		} else {
-			OriginType::Address
-		}
-	}
-}
-
-impl From<PalletAssetErr> for PSP22Error {
-    fn from(e: PalletAssetErr) -> PSP22Error {
-        match e {
-            PalletAssetErr::Other => PSP22Error::Custom(String::from("psp22 error")),
-            PalletAssetErr::CannotLookup => PSP22Error::Custom(String::from("CannotLookup")),
-            PalletAssetErr::BadOrigin => PSP22Error::Custom(String::from("BadOrigin")),
-            PalletAssetErr::Module => PSP22Error::Custom(String::from("Module")),
-            PalletAssetErr::ConsumerRemaining => PSP22Error::Custom(String::from("ConsumerRemaining")),
-            PalletAssetErr::NoProviders => PSP22Error::Custom(String::from("NoProviders")),
-            PalletAssetErr::TooManyConsumers => PSP22Error::Custom(String::from("TooManyConsumers")),
-            PalletAssetErr::Token(token_err) => PSP22Error::Custom(String::from("Token")),
-            PalletAssetErr::Arithmetic(arithmetic_error) => PSP22Error::Custom(String::from("Arithmetic")),
-            _ => PSP22Error::Custom(String::from("Unnown")),
-        }
-    }
-}
-
-impl ink_env::chain_extension::FromStatusCode for PalletAssetErr {
-    fn from_status_code(status_code: u32) -> Result<(), Self> {
-        match status_code {
-            0 => Ok(()),
-            _ => panic!("encountered unknown status code"),
-        }
-    }
-}
-
-impl From<scale::Error> for PalletAssetErr {
-    fn from(_: scale::Error) -> Self {
-        panic!("encountered unexpected invalid SCALE encoding")
-    }
-}
-
+use crate::pallet_assets::*;
 
 // #[brush::contract(env = crate::CustomEnvironment)]
 #[brush::contract]
